Create Watch-Presence.ps1
Watch for MS Teams Presence information and relay that to Home Assistant via a webhook. 

This enables automations such as changing the color of a WLED-driven light to let your family know if you are in a meeting.

The MS Teams log is tracked here, making this suitable for environments where the MS Graph API is restricted.